+++ src/main/java/com/internet/shop/controller/cart/AddProductToCartController.java
@@ -1,4 +1,4 @@
-package com.internet.shop.controllers.cart;
+package com.internet.shop.controller.cart;
 
 import com.internet.shop.lib.Injector;
 import com.internet.shop.model.Product;
@@ -6,12 +6,13 @@ import com.internet.shop.model.ShoppingCart;
 import com.internet.shop.service.ProductService;
 import com.internet.shop.service.ShoppingCartService;
 import java.io.IOException;
-import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class RemoveProductFromCartController extends HttpServlet {
+@WebServlet("/products/add-to-cart")
+public class AddProductToCartController extends HttpServlet {
     private static final Long USER_ID = 1L;
     private static final Injector INJECTOR = Injector.getInstance("com.internet.shop");
     private final ShoppingCartService shoppingCartService =
@@ -21,10 +22,12 @@ public class RemoveProductFromCartController extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
+            throws IOException {
+        String productId = req.getParameter("id");
         ShoppingCart shoppingCart = shoppingCartService.getByUserId(USER_ID);
-        Product product = productService.get(Long.valueOf(req.getParameter("id")));
-        shoppingCartService.deleteProduct(shoppingCart, product);
-        resp.sendRedirect(req.getContextPath() + "/cart/show");
+        Product product = productService.get(Long.valueOf(productId));
+        shoppingCartService.addProduct(shoppingCart, product);
+        resp.sendRedirect(req.getContextPath() + "/products/all");
+
     }
 }
