+++ src/main/java/com/internet/shop/controller/cart/CompleteOrderController.java
@@ -1,4 +1,4 @@
-package com.internet.shop.controllers.cart;
+package com.internet.shop.controller.cart;
 
 import com.internet.shop.lib.Injector;
 import com.internet.shop.model.Order;
@@ -10,10 +10,12 @@ import com.internet.shop.service.UserService;
 import java.io.IOException;
 import java.util.List;
 import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+@WebServlet("/orders/complete")
 public class CompleteOrderController extends HttpServlet {
     private static final Long USER_ID = 1L;
     private static final Injector INJECTOR = Injector.getInstance("com.internet.shop");
