+++ src/main/java/com/internet/shop/controller/cart/RemoveProductFromCartController.java
@@ -1,0 +1,31 @@
+package com.internet.shop.controller.cart;
+
+import com.internet.shop.lib.Injector;
+import com.internet.shop.model.Product;
+import com.internet.shop.model.ShoppingCart;
+import com.internet.shop.service.ProductService;
+import com.internet.shop.service.ShoppingCartService;
+import java.io.IOException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@WebServlet("/products/remove-from-cart")
+public class RemoveProductFromCartController extends HttpServlet {
+    private static final Long USER_ID = 1L;
+    private static final Injector INJECTOR = Injector.getInstance("com.internet.shop");
+    private final ShoppingCartService shoppingCartService =
+            (ShoppingCartService) INJECTOR.getInstance(ShoppingCartService.class);
+    private final ProductService productService =
+            (ProductService) INJECTOR.getInstance(ProductService.class);
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
+            throws IOException {
+        ShoppingCart shoppingCart = shoppingCartService.getByUserId(USER_ID);
+        Product product = productService.get(Long.valueOf(req.getParameter("id")));
+        shoppingCartService.deleteProduct(shoppingCart, product);
+        resp.sendRedirect(req.getContextPath() + "/cart/show");
+    }
+}
