+++ src/main/java/com/internet/shop/controller/cart/ShowShoppingCartController.java
@@ -1,32 +1,28 @@
-package com.internet.shop.controllers.cart;
+package com.internet.shop.controller.cart;
 
 import com.internet.shop.lib.Injector;
 import com.internet.shop.model.Product;
-import com.internet.shop.model.ShoppingCart;
-import com.internet.shop.service.ProductService;
 import com.internet.shop.service.ShoppingCartService;
 import java.io.IOException;
+import java.util.List;
 import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class AddProductToCartController extends HttpServlet {
+@WebServlet("/cart/show")
+public class ShowShoppingCartController extends HttpServlet {
     private static final Long USER_ID = 1L;
     private static final Injector INJECTOR = Injector.getInstance("com.internet.shop");
     private final ShoppingCartService shoppingCartService =
             (ShoppingCartService) INJECTOR.getInstance(ShoppingCartService.class);
-    private final ProductService productService =
-            (ProductService) INJECTOR.getInstance(ProductService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        String productId = req.getParameter("id");
-        ShoppingCart shoppingCart = shoppingCartService.getByUserId(USER_ID);
-        Product product = productService.get(Long.valueOf(productId));
-        shoppingCartService.addProduct(shoppingCart, product);
-        resp.sendRedirect(req.getContextPath() + "/products/all");
-
+        List<Product> products = shoppingCartService.getByUserId(USER_ID).getProducts();
+        req.setAttribute("products", products);
+        req.getRequestDispatcher("/WEB-INF/views/cart/show.jsp").forward(req, resp);
     }
 }
