+++ src/main/java/com/internet/shop/controller/orders/DeleteOrderController.java
@@ -1,13 +1,14 @@
-package com.internet.shop.controllers.orders;
+package com.internet.shop.controller.orders;
 
 import com.internet.shop.lib.Injector;
 import com.internet.shop.service.OrderService;
 import java.io.IOException;
-import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+@WebServlet("/orders/delete")
 public class DeleteOrderController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("com.internet.shop");
     private final OrderService orderService =
@@ -15,7 +16,7 @@ public class DeleteOrderController extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
+            throws IOException {
         orderService.delete(Long.valueOf(req.getParameter("id")));
         resp.sendRedirect(req.getContextPath() + "/orders/all");
     }
