+++ src/main/java/com/internet/shop/controller/orders/ShowOrderController.java
@@ -1,29 +1,26 @@
-package com.internet.shop.controllers.orders;
+package com.internet.shop.controller.orders;
 
 import com.internet.shop.lib.Injector;
 import com.internet.shop.model.Order;
 import com.internet.shop.service.OrderService;
-import com.internet.shop.service.UserService;
 import java.io.IOException;
-import java.util.List;
 import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class ShowAllOrdersController extends HttpServlet {
-    private static final Long USER_ID = 1L;
+@WebServlet("/orders/show")
+public class ShowOrderController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("com.internet.shop");
     private final OrderService orderService =
             (OrderService) INJECTOR.getInstance(OrderService.class);
-    private final UserService userService =
-            (UserService) INJECTOR.getInstance(UserService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<Order> orders = orderService.getUserOrders(userService.get(USER_ID));
-        req.setAttribute("orders", orders);
-        req.getRequestDispatcher("/WEB-INF/views/orders/all.jsp").forward(req, resp);
+        Order order = orderService.get(Long.valueOf(req.getParameter("id")));
+        req.setAttribute("order", order);
+        req.getRequestDispatcher("/WEB-INF/views/orders/show.jsp").forward(req, resp);
     }
 }
