+++ src/main/java/com/internet/shop/controller/products/CreateProductController.java
@@ -1,14 +1,16 @@
-package com.internet.shop.controllers.products;
+package com.internet.shop.controller.products;
 
 import com.internet.shop.lib.Injector;
 import com.internet.shop.model.Product;
 import com.internet.shop.service.ProductService;
 import java.io.IOException;
 import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+@WebServlet("/products/manage")
 public class CreateProductController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("com.internet.shop");
     private final ProductService productService =
