+++ src/main/java/com/internet/shop/controller/products/DeleteProductController.java
@@ -1,13 +1,14 @@
-package com.internet.shop.controllers.products;
+package com.internet.shop.controller.products;
 
 import com.internet.shop.lib.Injector;
 import com.internet.shop.service.ProductService;
 import java.io.IOException;
-import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+@WebServlet("/products/delete")
 public class DeleteProductController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("com.internet.shop");
     private final ProductService productService =
@@ -15,9 +16,9 @@ public class DeleteProductController extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
+            throws IOException {
         productService.delete(Long.valueOf(req.getParameter("id")));
         req.setAttribute("products", productService.getAll());
-        resp.sendRedirect(req.getContextPath() + "/products/new");
+        resp.sendRedirect(req.getContextPath() + "/products/manage");
     }
 }
