+++ src/main/java/com/internet/shop/controller/products/ShowAllProductsController.java
@@ -1,26 +1,27 @@
-package com.internet.shop.controllers.cart;
+package com.internet.shop.controller.products;
 
 import com.internet.shop.lib.Injector;
 import com.internet.shop.model.Product;
-import com.internet.shop.service.ShoppingCartService;
+import com.internet.shop.service.ProductService;
 import java.io.IOException;
 import java.util.List;
 import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class ShowShoppingCartController extends HttpServlet {
-    private static final Long USER_ID = 1L;
+@WebServlet("/products/all")
+public class ShowAllProductsController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("com.internet.shop");
-    private final ShoppingCartService shoppingCartService =
-            (ShoppingCartService) INJECTOR.getInstance(ShoppingCartService.class);
+    private final ProductService productService =
+            (ProductService) INJECTOR.getInstance(ProductService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<Product> products = shoppingCartService.getByUserId(USER_ID).getProducts();
+        List<Product> products = productService.getAll();
         req.setAttribute("products", products);
-        req.getRequestDispatcher("/WEB-INF/views/cart/show.jsp").forward(req, resp);
+        req.getRequestDispatcher("/WEB-INF/views/products/all.jsp").forward(req, resp);
     }
 }
