+++ src/main/java/com/internet/shop/controller/users/DeleteUserController.java
@@ -1,25 +1,24 @@
-package com.internet.shop.controllers.users;
+package com.internet.shop.controller.users;
 
 import com.internet.shop.lib.Injector;
-import com.internet.shop.model.User;
 import com.internet.shop.service.UserService;
 import java.io.IOException;
-import java.util.List;
-import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class ShowAllUsersController extends HttpServlet {
+@WebServlet("/users/delete")
+public class DeleteUserController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("com.internet.shop");
     private final UserService userService =
             (UserService) INJECTOR.getInstance(UserService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        List<User> users = userService.getAll();
-        req.setAttribute("users", users);
-        req.getRequestDispatcher("/WEB-INF/views/users/all.jsp").forward(req, resp);
+            throws IOException {
+        String userId = req.getParameter("id");
+        userService.delete(Long.valueOf(userId));
+        resp.sendRedirect(req.getContextPath() + "/users/all");
     }
 }
