+++ src/main/java/com/internet/shop/controller/users/ShowAllUsersController.java
@@ -1,14 +1,18 @@
-package com.internet.shop.controllers.users;
+package com.internet.shop.controller.users;
 
 import com.internet.shop.lib.Injector;
+import com.internet.shop.model.User;
 import com.internet.shop.service.UserService;
 import java.io.IOException;
+import java.util.List;
 import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class DeleteUserController extends HttpServlet {
+@WebServlet("/users/all")
+public class ShowAllUsersController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("com.internet.shop");
     private final UserService userService =
             (UserService) INJECTOR.getInstance(UserService.class);
@@ -16,8 +20,8 @@ public class DeleteUserController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        String userId = req.getParameter("id");
-        userService.delete(Long.valueOf(userId));
-        resp.sendRedirect(req.getContextPath() + "/users/all");
+        List<User> users = userService.getAll();
+        req.setAttribute("users", users);
+        req.getRequestDispatcher("/WEB-INF/views/users/all.jsp").forward(req, resp);
     }
 }
