--- src/main/java/com/internet/shop/controllers/products/ShowAllProductsController.java
@@ -1,25 +1,0 @@
-package com.internet.shop.controllers.products;
-
-import com.internet.shop.lib.Injector;
-import com.internet.shop.model.Product;
-import com.internet.shop.service.ProductService;
-import java.io.IOException;
-import java.util.List;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public class ShowAllProductsController extends HttpServlet {
-    private static final Injector INJECTOR = Injector.getInstance("com.internet.shop");
-    private final ProductService productService =
-            (ProductService) INJECTOR.getInstance(ProductService.class);
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        List<Product> products = productService.getAll();
-        req.setAttribute("products", products);
-        req.getRequestDispatcher("/WEB-INF/views/products/all.jsp").forward(req, resp);
-    }
-}
