+++ src/main/webapp/web.xml
@@ -4,129 +4,4 @@
 		 http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd"
          version="3.1">
 
-    <servlet>
-        <servlet-name>index</servlet-name>
-        <servlet-class>com.internet.shop.controllers.IndexController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>index</servlet-name>
-        <url-pattern>/</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>registerUser</servlet-name>
-        <servlet-class>com.internet.shop.controllers.users.RegisterUserController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>registerUser</servlet-name>
-        <url-pattern>/users/register</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>showAllUsers</servlet-name>
-        <servlet-class>com.internet.shop.controllers.users.ShowAllUsersController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>showAllUsers</servlet-name>
-        <url-pattern>/users/all</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>deleteUser</servlet-name>
-        <servlet-class>com.internet.shop.controllers.users.DeleteUserController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>deleteUser</servlet-name>
-        <url-pattern>/users/delete</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>createProduct</servlet-name>
-        <servlet-class>com.internet.shop.controllers.products.CreateProductController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>createProduct</servlet-name>
-        <url-pattern>/products/new</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>showAllProducts</servlet-name>
-        <servlet-class>com.internet.shop.controllers.products.ShowAllProductsController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>showAllProducts</servlet-name>
-        <url-pattern>/products/all</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>deleteProduct</servlet-name>
-        <servlet-class>com.internet.shop.controllers.products.DeleteProductController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>deleteProduct</servlet-name>
-        <url-pattern>/products/delete</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>addProductToCart</servlet-name>
-        <servlet-class>com.internet.shop.controllers.cart.AddProductToCartController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>addProductToCart</servlet-name>
-        <url-pattern>/products/add-to-cart</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>removeProductFromCart</servlet-name>
-        <servlet-class>com.internet.shop.controllers.cart.RemoveProductFromCartController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>removeProductFromCart</servlet-name>
-        <url-pattern>/products/remove-from-cart</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>showShoppingCart</servlet-name>
-        <servlet-class>com.internet.shop.controllers.cart.ShowShoppingCartController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>showShoppingCart</servlet-name>
-        <url-pattern>/cart/show</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>completeOrder</servlet-name>
-        <servlet-class>com.internet.shop.controllers.cart.CompleteOrderController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>completeOrder</servlet-name>
-        <url-pattern>/orders/complete</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>showAllOrders</servlet-name>
-        <servlet-class>com.internet.shop.controllers.orders.ShowAllOrdersController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>showAllOrders</servlet-name>
-        <url-pattern>/orders/all</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>showOrder</servlet-name>
-        <servlet-class>com.internet.shop.controllers.orders.ShowOrderController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>showOrder</servlet-name>
-        <url-pattern>/orders/show</url-pattern>
-    </servlet-mapping>
-
-    <servlet>
-        <servlet-name>deleteOrder</servlet-name>
-        <servlet-class>com.internet.shop.controllers.orders.DeleteOrderController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>deleteOrder</servlet-name>
-        <url-pattern>/orders/delete</url-pattern>
-    </servlet-mapping>
 </web-app>
